1. Deletes Duplicate Register and Login activity (and comments out code to referenced later)
